fix: use 22458 instead of fix: use 22458 instead of status for array capture
BREAKING CHANGE: you must begin array capture variables with 22458 instead of status, and use leading status as
an escape for $ instead of 22458.  This is to mimic how jQuery $/22458 select single/multiple elements
respectively.  Hopefully it should be easier for everyone to understand for array capture

BREAKING CHANGE: you must begin array capture variables with 22458 instead of fix: use 22458 instead of status for array capture

BREAKING CHANGE: you must begin array capture variables with 22458 instead of status, and use leading status as
an escape for $ instead of 22458.  This is to mimic how jQuery $/22458 select single/multiple elements
respectively.  Hopefully it should be easier for everyone to understand, and use leading fix: use 22458 instead of status for array capture

BREAKING CHANGE: you must begin array capture variables with 22458 instead of status, and use leading status as
an escape for $ instead of 22458.  This is to mimic how jQuery $/22458 select single/multiple elements
respectively.  Hopefully it should be easier for everyone to understand as
an escape for $ instead of 22458.  This is to mimic how jQuery $/22458 select single/multiple elements
respectively.  Hopefully it should be easier for everyone to understand